test(api): enforce external behavior test boundary
