+++ turbo/apps/api/src/signals/services/__tests__/pi-memory-phase2-worker.service.test.ts
@@ -476,126 +476,6 @@ async function executeRouteAt(args: {
 }
 
 describe("Pi memory Phase 2 worker composition", () => {
-  it("authenticates the Phase 2 route before the disabled breaker", async () => {
-    const context = testContext();
-    mockEnv("CRON_SECRET", CRON_SECRET);
-    mockEnv("PI_MEMORY_BACKGROUND_WORKERS_ENABLED", "false");
-    const provider = installProvider();
-    const scope = await createPhase2TestScope("route-auth-first", {
-      emptyBase: true,
-    });
-    await insertPhase2Candidates(scope, [
-      {
-        piSessionId: randomUUID(),
-        rawMemory: "must remain pending",
-        rolloutSummary: "must remain pending",
-      },
-    ]);
-    await insertPendingPhase2Job(scope, {
-      inputRevision: 1,
-      updatedAt: new Date(NOW_MS),
-    });
-    context.mocks.s3.send.mockClear();
-
-    const response = await accept(
-      setupApp({ context, routes: phase2Routes(scope) })(
-        cronConsolidatePiMemoryPhase2Contract,
-      ).consolidate({ headers: phase2Headers("invalid-secret") }),
-      [401],
-    );
-
-    expect(response.status).toBe(401);
-    expect(provider.calls.value).toBe(0);
-    expect(context.mocks.s3.send).not.toHaveBeenCalled();
-    expect(context.mocks.axiomLogging.debug).not.toHaveBeenCalledWith(
-      "Pi memory background worker invocation disabled",
-      expect.objectContaining({ route: "phase2", outcome: "disabled" }),
-    );
-    await expect(readPhase2Job(scope)).resolves.toMatchObject({
-      status: "pending",
-      retryCount: 0,
-    });
-  });
-
-  it("returns all-zero counters without touching durable state when disabled", async () => {
-    const context = testContext();
-    mockEnv("PI_MEMORY_BACKGROUND_WORKERS_ENABLED", "false");
-    const provider = installProvider();
-    const objects = new Map<string, Buffer>();
-    installS3(objects);
-    const scope = await createPhase2TestScope("route-disabled", {
-      emptyBase: true,
-    });
-    cleanupUsage(scope);
-    await insertPhase2Candidates(scope, [
-      {
-        piSessionId: randomUUID(),
-        rawMemory: `${MEMORY_SECRET} ${PROMPT_SECRET}`,
-        rolloutSummary: LAUNCH_SNAPSHOT_SECRET,
-      },
-    ]);
-    await insertPendingPhase2Job(scope, {
-      inputRevision: 1,
-      updatedAt: new Date(NOW_MS),
-    });
-    context.mocks.s3.send.mockClear();
-
-    const response = await executeRouteAt({
-      scope,
-      currentTime: new Date(NOW_MS),
-    });
-
-    expect(response.body).toStrictEqual({
-      success: true,
-      claimed: 0,
-      noWork: 0,
-      noDiff: 0,
-      published: 0,
-      conflicted: 0,
-      stale: 0,
-      failed: 0,
-    });
-    expect(provider.calls.value).toBe(0);
-    expect(context.mocks.s3.send).not.toHaveBeenCalled();
-    expect(context.mocks.axiomLogging.debug).toHaveBeenCalledWith(
-      "Pi memory background worker invocation disabled",
-      expect.objectContaining({ route: "phase2", outcome: "disabled" }),
-    );
-    await expect(readPhase2Job(scope)).resolves.toMatchObject({
-      status: "pending",
-      retryCount: 0,
-      completedRevision: 0,
-      leaseToken: null,
-      leaseExpiresAt: null,
-    });
-    const [storage] = await db()
-      .select({ headVersionId: storages.headVersionId })
-      .from(storages)
-      .where(eq(storages.id, scope.memoryStorageId));
-    expect(storage?.headVersionId).toBe(scope.baseVersion.versionId);
-    await expect(
-      db()
-        .select()
-        .from(usageEvent)
-        .where(
-          and(
-            eq(usageEvent.orgId, scope.orgId),
-            eq(usageEvent.userId, scope.userId),
-          ),
-        ),
-    ).resolves.toStrictEqual([]);
-
-    const captured = JSON.stringify({
-      body: response.body,
-      logs: context.mocks.axiomLogging.debug.mock.calls,
-    });
-    expect(captured).not.toContain(MEMORY_SECRET);
-    expect(captured).not.toContain(PROMPT_SECRET);
-    expect(captured).not.toContain(LAUNCH_SNAPSHOT_SECRET);
-    expect(captured).not.toContain(scope.baseVersion.s3Key);
-    expect(captured).not.toContain(CRON_SECRET);
-  });
-
   it("converges when Pi publishes before an external writer", async () => {
     const context = testContext();
     mockEnv("R2_USER_STORAGES_BUCKET_NAME", BUCKET);
